requirements.txt
asgiref==3.8.1
Django==5.0.1
djangorestframework==3.14.0
psycopg2-binary==2.9.10
pytz==2025.2
sqlparse==0.5.3
psycopg2-binary==2.9.10
drf-spectacular==0.28.0
Pillow==11.2.1